requirements.txt
asgiref==3.6.0
Django==4.2
gunicorn==20.1.0
Pillow==9.5.0
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.4.0
